Move testing classes from org.apache.maven.api to org.apache.maven.testing
Move classes from org.apache.maven.api.di.testing and
org.apache.maven.api.plugin.testing to org.apache.maven.testing.di
and org.apache.maven.testing.plugin
