+++ impl/maven-testing/src/main/java/org/apache/maven/testing/di/MavenDIExtension.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.api.di.testing;
+package org.apache.maven.testing.di;
 
 import java.io.File;
 import java.util.Optional;
