+++ impl/maven-testing/src/main/java/org/apache/maven/testing/di/MavenDITest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.api.di.testing;
+package org.apache.maven.testing.di;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
