+++ impl/maven-testing/src/main/java/org/apache/maven/testing/plugin/stubs/LookupStub.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.api.plugin.testing.stubs;
+package org.apache.maven.testing.plugin.stubs;
 
 import java.util.List;
 import java.util.Map;
