+++ its/core-it-suite/src/test/resources/gh-11055-di-service-injection/src/test/java/com/gitlab/tkslaw/ditests/InjectServiceMojoTests.java
@@ -18,8 +18,8 @@
  */
 package com.gitlab.tkslaw.ditests;
 
-import org.apache.maven.api.plugin.testing.InjectMojo;
-import org.apache.maven.api.plugin.testing.MojoTest;
+import org.apache.maven.testing.plugin.InjectMojo;
+import org.apache.maven.testing.plugin.MojoTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
