+++ its/core-it-suite/src/test/resources/gh-11055-di-service-injection/src/test/java/com/gitlab/tkslaw/ditests/TestProviders.java
@@ -24,7 +24,7 @@ import org.apache.maven.api.di.Named;
 import org.apache.maven.api.di.Priority;
 import org.apache.maven.api.di.Provides;
 import org.apache.maven.api.di.Singleton;
-import org.apache.maven.api.plugin.testing.stubs.SessionMock;
+import org.apache.maven.testing.plugin.stubs.SessionMock;
 import org.apache.maven.api.services.DependencyResolver;
 import org.apache.maven.api.services.OsService;
 import org.apache.maven.api.services.ToolchainManager;
@@ -33,7 +33,7 @@ import org.apache.maven.impl.InternalSession;
 import org.apache.maven.impl.model.DefaultOsService;
 import org.mockito.Mockito;
 
-import static org.apache.maven.api.plugin.testing.MojoExtension.getBasedir;
+import static org.apache.maven.testing.plugin.MojoExtension.getBasedir;
 
 @Named
 public class TestProviders {
